Updated TypeGenerator and TestSuiteGenerator
The TypeGenerator now enumerates the full (irregular) type space, rather
than the regularised extension approach it took previously.

The TestSuiteGenerator now determines whether or not one type should
subtype another using a semantic interpretation of types.  This
contrasts with the old approach which was based on the (broken) existing
Whiley subtype operator.